Don't stop TPUs which die due to failure.

+++ bin/bake_benchmark.py
@@ -141,13 +141,16 @@ BENCHMARK_EXIT_CODE=$?
 
 set -e
 
-echo  gcloud alpha compute tpus delete $MLP_TPU_NAME --zone $MLP_GCP_ZONE
-yes | gcloud alpha compute tpus delete $MLP_TPU_NAME --zone $MLP_GCP_ZONE
+if [ $? -ne 0 ]
+  then
+    echo  gcloud alpha compute tpus delete $MLP_TPU_NAME --zone $MLP_GCP_ZONE
+    yes | gcloud alpha compute tpus delete $MLP_TPU_NAME --zone $MLP_GCP_ZONE
 
-# Stop the TPU Sidecar
-if [[ $MLP_TPU_SIDECAR =~ "Y"$ ]]; then
-    echo  gcloud alpha compute tpus delete $MLP_TPU_SIDECAR_NAME --zone $MLP_GCP_ZONE
-    yes | gcloud alpha compute tpus delete $MLP_TPU_SIDECAR_NAME --zone $MLP_GCP_ZONE
+    # Stop the TPU Sidecar
+    if [[ $MLP_TPU_SIDECAR =~ "Y"$ ]]; then
+        echo  gcloud alpha compute tpus delete $MLP_TPU_SIDECAR_NAME --zone $MLP_GCP_ZONE
+        yes | gcloud alpha compute tpus delete $MLP_TPU_SIDECAR_NAME --zone $MLP_GCP_ZONE
+    fi
 fi
 
 exit $BENCHMARK_EXIT_CODE
